+++ apps/alpha-test-app/src/App.vue
@@ -1,33 +1,46 @@
 <script setup lang="ts">
 import logout from "@sit-onyx/icons/logout.svg?raw";
 import settings from "@sit-onyx/icons/settings.svg?raw";
+import { useDark, useToggle } from "@vueuse/core";
 import {
   OnyxAppLayout,
   OnyxNavBar,
   OnyxNavItem,
+  OnyxSwitch,
   OnyxUserMenu,
   type ListboxOption,
   type OnyxNavItemProps,
 } from "sit-onyx";
 import { RouterView, useRouter } from "vue-router";
 import onyxLogo from "./assets/onyx-logo.svg";
+import { useGridStore } from "./stores/grid-store";
 
 const router = useRouter();
+const gridStore = useGridStore();
 
 const navItems = [
   { label: "Home", href: "/" },
   { label: "Form Demo", href: "/form-demo" },
   { label: "Layout Demo", href: "/layout-demo" },
+  { label: "Grid Demo", href: "/grid" },
 ] satisfies OnyxNavItemProps[];
 
 const userMenuOptions = [
   { value: "/settings", label: "Settings", icon: settings },
   { value: "logout", label: "Logout", icon: logout, color: "danger" },
 ] satisfies ListboxOption[];
+
+const isDark = useDark();
+const toggleDark = useToggle(isDark);
 </script>
 
 <template>
-  <OnyxAppLayout>
+  <OnyxAppLayout
+    :class="{
+      'onyx-grid-max-md': gridStore.isMaxWidth,
+      'onyx-grid-center': gridStore.isCentered,
+    }"
+  >
     <template #navBar>
       <OnyxNavBar
         app-name="Alpha Test App"
@@ -45,6 +58,9 @@ const userMenuOptions = [
         />
 
         <template #contextArea>
+          <!-- TODO: include the theme selection into the user menu once it supports that feature -->
+          <OnyxSwitch label="Dark Mode" :model-value="isDark" @update:model-value="toggleDark" />
+
           <OnyxUserMenu username="John Doe" :options="userMenuOptions">
             <template #footer>
               App Version