+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,19 @@
 # playground
 
+## 0.0.1-alpha.44
+
+### Patch Changes
+
+- Updated dependencies [c011e27]
+  - sit-onyx@1.0.0-alpha.105
+
+## 0.0.1-alpha.43
+
+### Patch Changes
+
+- Updated dependencies [20fb878]
+  - sit-onyx@1.0.0-alpha.104
+
 ## 0.0.1-alpha.42
 
 ### Patch Changes